requirements.txt
torch
torchvision
numpy
gradio
Pillow
scikit-learn
tqdm
matplotli